requirements.txt
Pillow
pyautogui
pyperclip
anthropic
python-dotenv
